moving broker outside of the gateway sub-dir
