+++ internal/ollama/client.go
@@ -32,102 +32,6 @@ func NewClient(baseURL string, log *config.Logger) *Client {
 	}
 }
 
-// Generate sends a single-turn prompt to Ollama's /api/generate endpoint.
-// Deprecated: Use Chat instead. This endpoint lacks tool-calling support.
-func (c *Client) Generate(ctx context.Context, req Request, streamCallback func(string)) (*Response, error) {
-	endpoint := fmt.Sprintf("%s/api/generate", c.BaseURL)
-
-	ollamaReq := GenerateRequest{
-		Model:  req.Model,
-		Prompt: req.Prompt,
-		System: req.System,
-		Format: req.Format,
-		Stream: req.Stream,
-	}
-
-	payloadBytes, err := json.Marshal(ollamaReq)
-	if err != nil {
-		return nil, fmt.Errorf("failed to marshal request: %w", err)
-	}
-
-	httpReq, err := http.NewRequestWithContext(ctx, http.MethodPost, endpoint, bytes.NewReader(payloadBytes))
-	if err != nil {
-		return nil, fmt.Errorf("failed to create request: %w", err)
-	}
-	httpReq.Header.Set("Content-Type", "application/json")
-
-	resp, err := c.HTTPClient.Do(httpReq)
-	if err != nil {
-		return nil, fmt.Errorf("Ollama API request failed: %w", err)
-	}
-	defer resp.Body.Close()
-
-	var finalResponse Response
-	finalResponse.Model = req.Model
-
-	if req.Stream && streamCallback != nil {
-		scanner := bufio.NewScanner(resp.Body)
-		for scanner.Scan() {
-			var chunk GenerateResponse
-			if err := json.Unmarshal(scanner.Bytes(), &chunk); err != nil {
-				c.log.Debug("Ollama stream: failed to parse chunk: %v | raw: %q", err, scanner.Text())
-				continue
-			}
-
-			streamCallback(chunk.Response)
-			finalResponse.Response += chunk.Response
-			finalResponse.Thinking += chunk.Thinking
-
-			if chunk.Done {
-				finalResponse.TotalDuration = chunk.TotalDuration
-				finalResponse.EvalDuration = chunk.EvalDuration
-				finalResponse.EvalCount = chunk.EvalCount
-				finalResponse.PromptEvalDuration = chunk.PromptEvalDuration
-			}
-		}
-		if err := scanner.Err(); err != nil {
-			c.log.Warn("Ollama stream: scanner error: %v", err)
-		}
-
-		if finalResponse.Response == "" && finalResponse.Thinking != "" {
-			finalResponse.Response = finalResponse.Thinking
-		}
-		return &finalResponse, nil
-	}
-
-	rawBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read response body: %w", err)
-	}
-
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		c.log.Error("Ollama returned HTTP %d: %s", resp.StatusCode, string(rawBody))
-		return nil, fmt.Errorf("Ollama returned HTTP %d", resp.StatusCode)
-	}
-
-	var ollamaResp GenerateResponse
-	if err := json.Unmarshal(rawBody, &ollamaResp); err != nil {
-		c.log.Error("Ollama response decode failed: %v | raw: %q", err, string(rawBody))
-		return nil, fmt.Errorf("failed to decode response: %w", err)
-	}
-
-	response := ollamaResp.Response
-	if response == "" && ollamaResp.Thinking != "" {
-		response = ollamaResp.Thinking
-	}
-
-	return &Response{
-		Model:              ollamaResp.Model,
-		Response:           response,
-		Thinking:           ollamaResp.Thinking,
-		TotalDuration:      ollamaResp.TotalDuration,
-		LoadDuration:       ollamaResp.LoadDuration,
-		PromptEvalDuration: ollamaResp.PromptEvalDuration,
-		EvalDuration:       ollamaResp.EvalDuration,
-		EvalCount:          ollamaResp.EvalCount,
-	}, nil
-}
-
 // Show retrieves model metadata from Ollama's /api/show endpoint.
 func (c *Client) Show(ctx context.Context, req ShowRequest) (*ShowResponse, error) {
 	endpoint := fmt.Sprintf("%s/api/show", c.BaseURL)
